auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arrow>=1.3.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.11",
    "jwt>=1.3.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
    "types-greenlet>=3.1.0.20250318",
    "types-passlib>=1.7.7.20250326",
    "types-psycopg2>=2.9.21.20250318",
    "types-pyasn1>=0.6.0.20250208",
    "types-python-dateutil>=2.9.0.20241206",
    "types-python-jose>=3.4.0.20250224",
    "types-ujson>=5.10.0.20250326",
    "uvicorn>=0.34.0",
]
